fix: [CDS-34063]: more covrage
