+++ src/modules/72-templates-library/components/TemplateStudio/PipelineTemplateCanvas/PipelineTemplateCanvas.tsx
@@ -1,30 +1,28 @@
 /*
- * Copyright 2021 Harness Inc. All rights reserved.
+ * Copyright 2022 Harness Inc. All rights reserved.
  * Use of this source code is governed by the PolyForm Shield 1.0.0 license
  * that can be found in the licenses directory at the root of this repository, also available at
  * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
  */
 
-import produce from 'immer'
-import { get, isEmpty, isUndefined, omit, omitBy, set } from 'lodash-es'
 import React from 'react'
 import { useParams } from 'react-router-dom'
-import {
-  DefaultNewStageId,
-  DefaultNewStageName
-} from '@templates-library/components/TemplateStudio/StageTemplateCanvas/StageTemplateForm/StageTemplateForm'
-import { TemplatePipelineProvider } from '@pipeline/components/TemplatePipelineContext'
-import { StageTemplateCanvasWithRef } from '@templates-library/components/TemplateStudio/StageTemplateCanvas/StageTemplateCanvas'
-import type { PipelineInfoConfig } from 'services/cd-ng'
+import { omit, set } from 'lodash-es'
 import { TemplateContext } from '@templates-library/components/TemplateStudio/TemplateContext/TemplateContext'
-import type { TemplateFormRef } from '@templates-library/components/TemplateStudio/TemplateStudio'
-import { DefaultPipeline } from '@pipeline/components/PipelineStudio/PipelineContext/PipelineActions'
 import type { GitQueryParams, ProjectPathProps } from '@common/interfaces/RouteInterfaces'
-import { sanitize } from '@common/utils/JSONUtils'
 import { useQueryParams } from '@common/hooks'
+import { TemplatePipelineProvider } from '@pipeline/components/TemplatePipelineContext'
+import { sanitize } from '@common/utils/JSONUtils'
+import type { PipelineInfoConfig } from 'services/cd-ng'
+import StageBuilder from '@pipeline/components/PipelineStudio/StageBuilder/StageBuilder'
 import { PipelineContextType } from '@pipeline/components/PipelineStudio/PipelineContext/PipelineContext'
+import { TemplateDrawer } from '@templates-library/components/TemplateDrawer/TemplateDrawer'
+import { RightDrawer } from '@pipeline/components/PipelineStudio/RightDrawer/RightDrawer'
 
-const StageTemplateCanvasWrapper = (_props: unknown, formikRef: TemplateFormRef) => {
+export const DefaultNewPipelineName = 'Pipeline Name'
+export const DefaultNewPipelineId = 'pipeline_name'
+
+const PipelineTemplateCanvas = () => {
   const {
     state: { template, isLoading, isUpdated },
     updateTemplate,
@@ -34,12 +32,7 @@ const StageTemplateCanvasWrapper = (_props: unknown, formikRef: TemplateFormRef)
   const { branch, repoIdentifier } = useQueryParams<GitQueryParams>()
 
   const createPipelineFromTemplate = React.useCallback(
-    () =>
-      produce({ ...DefaultPipeline }, draft => {
-        if (!isEmpty(template.spec)) {
-          set(draft, 'stages[0].stage', { ...template.spec, name: DefaultNewStageName, identifier: DefaultNewStageId })
-        }
-      }),
+    () => ({ ...template.spec, name: DefaultNewPipelineName, identifier: DefaultNewPipelineId }),
     [template.spec]
   )
 
@@ -52,8 +45,7 @@ const StageTemplateCanvasWrapper = (_props: unknown, formikRef: TemplateFormRef)
   }, [isLoading, isUpdated])
 
   const onUpdatePipeline = async (pipelineConfig: PipelineInfoConfig) => {
-    const stage = omitBy(omitBy(get(pipelineConfig, 'stages[0].stage'), isUndefined), isEmpty)
-    const processNode = omit(stage, 'name', 'identifier', 'description', 'tags')
+    const processNode = omit(pipelineConfig, 'name', 'identifier', 'description', 'tags')
     sanitize(processNode, { removeEmptyArray: false, removeEmptyObject: false, removeEmptyString: false })
     set(template, 'spec', processNode)
     await updateTemplate(template)
@@ -65,10 +57,12 @@ const StageTemplateCanvasWrapper = (_props: unknown, formikRef: TemplateFormRef)
         queryParams={{ accountIdentifier: accountId, orgIdentifier, projectIdentifier, repoIdentifier, branch }}
         initialValue={pipeline as PipelineInfoConfig}
         onUpdatePipeline={onUpdatePipeline}
-        contextType={PipelineContextType.StageTemplate}
+        contextType={PipelineContextType.PipelineTemplate}
         isReadOnly={isReadonly}
       >
-        <StageTemplateCanvasWithRef ref={formikRef} />
+        <StageBuilder />
+        <RightDrawer />
+        <TemplateDrawer />
       </TemplatePipelineProvider>
     )
   } else {
@@ -76,4 +70,4 @@ const StageTemplateCanvasWrapper = (_props: unknown, formikRef: TemplateFormRef)
   }
 }
 
-export const StageTemplateCanvasWrapperWithRef = React.forwardRef(StageTemplateCanvasWrapper)
+export const PipelineTemplateCanvasWithRef = React.forwardRef(PipelineTemplateCanvas)
